Timezone won't be correctly set for bogus MCC
Timezone is not updated if MCC is bogus like 001. If MCC is bogus, there
is no value for bogus MCC in MccTable, so, country iso will be empty. If
country iso is empty, NitzStateMachine#handleNetworkCountryCodeSet is
not called.

Timezone should be updated even if mcc is valid and country iso is
empty.

Test: atest LocaleTrackerTest
Bug: 142840879
Change-Id: Idf292aa2d9e3f3ae5678818a804d8cbfa03af076