Handle cell broadcasts with new module
This is gated by a flag sEnableCbModule. Also add logic to trigger cell
broadcast with an intent, gated with the flag TEST_MODE. The framework
interfaces with the module through CellBroadcastServiceManager, which
manages a single binding for all phones in mSIM cases.

Also add a dump(), which can be triggered with `adb shell dumpsys isms`

Bug: 135956699
Test: manual
Change-Id: I5b17e3ea010bab788b26f4c08539a33f38cbb20e