A new NitzStateMachine implementation and tests
A new NitzStateMachine implementation and tests. It is intended as a
drop-in replacement for NitzStateMachineImpl and the behavior should be
broadly the same. The implementation is deliberately decomposed to allow
for easier testing.

The new implementation is *not* enabled in this commit.

Bug:140712361
Test: atest com.android.internal.telephony.nitz
Merged-In: I02a96b2c7ea646b548bc874bdadcbffa0caa9ab8
Change-Id: I02a96b2c7ea646b548bc874bdadcbffa0caa9ab8