Modify tz detection logic on uninitialized devices
Make the tz detection more aggressive when the device's time zone
setting is uninitialized.

It is possible that bugs 141446564 & 141727033 were noticed as a result
of a behavior change between P and Q. Commit 325bce493bce1
removed use of isTimeZoneSettingInitialized in one flow, which might
have had the side effect of setting the time zone on uninitialized
devices.

This commit adds explicit handling for devices with uninitialized time
zone settings. The impact should be limited because the paths should
only be followed once.

Bug: 141727033
Bug: 141740972
Bug: 141446564
Test: atest com.android.internal.telephony.NitzStateMachineImplTest
Change-Id: I84202c8db90f576e9d8ef9d6285570cdaa147029