Remove getCachedNitzData() from NitzStateMachine
Remove getCachedNitzData() from the NitzStateMachine interface because
it is only used for a CDMA code path that probably won't exist much
longer. This simplifies the contract around NitzStateMachine. The new
behavior is not quite the same as the old (since NitzStateMachineImpl
does some validation, filtering, etc.) but it should be good enough for
the CDMA codepath.

Test: Treehugger
Bug: 140712361
Change-Id: Id083e27a140a55f15802e61f66c71929fb1c389c