Fix time zone offset calculation in ResponseData
Fix the time zone offset calculation in ResponseData: previously it
wasn't taking into account the current time and whether DST was in
force so would only be correct in countries without DST and countries
that do use DST during periods of DST.

This bug was noticed via inspection only. ResponseData may not be used
by anything.

Test: None
Change-Id: I412c41fe97217c1d48c70cc5ac589e0d96c5c632