Remove uninitialized tz path from MccTable
Remove code from MccTable that can set a device's time zone if the time
zone setting is not yet initialized but there is a known MCC. e.g. on
first boot when a SIM is present.

This is being done for a few reasons:
1) The telephony code responsible for setting time zone is
NitzStateMachine, which is scoped to a given phone. All telephony logic
that can set time zone should ideally be in one place.

2) There is logic in NitzStateMachineImpl that already handles setting
the time zone when the device is not initialized. It's unclear which
would win, but they should do the same thing and that makes the MccTable
code redundant.

Test: manual testing in AOSP with a SIM after factory reset
Bug: 142864894
Bug: 140712361
Change-Id: I4ec60ba921dc4442c6dbbcfb17d8aadeb36f1ae3