Add NitzStateMachine.handleNetworkUnavailable()
Indicate to NitzStateMachine when the network is unavailable so it can
clear network-device (i.e. NITZ) data. This is important now that the
LocaleTracker can now get a country without joining a network (this
changed in Q): the NITZ data could become very old, which leads to
incorrect time zone detection results when traveling.

Bug: 142979089
Test: atest com.android.internal.telephony.NitzStateMachineImplTest
Change-Id: I719532e44f49525800fd708c419b2d3d6d64d159