Pass country to NitzStateMachine
Have LocaleTracker explicitly pass country ISO code to NitzStateMachine.

Previously NitzStateMachine would request the country from the Phone
object. This change assumes that the LocaleTracker is now the authority
on the country to use for time zone detection. This will make it easier
to potentially supply a different country to the NitzStateMachine than
just the one set on the Phone object, i.e. if there's a distinction
between geographical country that the time zone detection needs and
country used for other reasons (e.g. the use of MCC 310, the US MCC in
Guam).

Bug: 143451294
Bug: 117863016
Test: atest com.android.internal.telephony.nitz
Test: atest com.android.internal.telephony.NitzStateMachineImplTest
Test: atest com.android.internal.telephony.LocaleTrackerTest
Change-Id: I5c5615c07c2a1bf8e11442513452fc03c3d900a3