Migrate ECM cancelled status from IMS->CS upon SRVCC
When there is an SRVCC for an emergency call after it has already
canceled the ECM notification, the "ecm cancelled" state was not
transferred from ImsPhone->GsmCdmaCallTracker. The ECM service would
still think it was in an emergency call and never leave that state
until a normal call was placed.

This change moves that state to Phone and migrates it over when
an SRVCC occurs.

Bug: 140259625
Test: manual verification (see bug for repro steps)
Merged-In: I7db732102471d201f17098fa8ae205624c1a64f9
Change-Id: I7db732102471d201f17098fa8ae205624c1a64f9