Move empty time suggestion logic to the server
Move empty time suggestion logic to the time detector service.
Currently there is no support for "withdrawing" a time suggestion in the
service. This change moves the stubbing to the server. Currently the
service just uses the latest suggestion it has received regardless of
origin. This presents a possible issue on multi-phone devices because
the device could alternate between two times as NITZ signals come in.

In future, when the service tracks the origin of the suggestions and
tries to be "sticky", this will allow the service to switch from one
origin to another if an origin becomes "unsure". e.g. if we kept using
the last signal from the phone with the lowest ID, and stopped receiving
NITZ updates to that phone, we would keep using the last suggestion
"forever". By withdrawing the suggestion from a phone when leaving the
network the service can switch to existing signals received from another
device.

Bug: 140712361
Test: atest com.android.internal.telephony.NitzStateMachineImplTest
Test: atest com.android.internal.telephony.nitz
Change-Id: I2fea29a582b8bf56adce931a12189864fbd23a13