Remove broadcast of NETWORK_SET_TIMEZONE intent
android.intent.action.NETWORK_SET_TIMEZONE is not used by anything in
the platform so it can be removed. It was previously broadcast when
telephony made a time zone change, presumably as a way to coordinate
time zone detection logic if there were other, lower-priority detection
mechanisms on the device.

All the APK usages appear to be only manifest entries for
<protected-broadcast> in apps that don't send the intent themselves,
which doesn't make a lot of sense.

Note: There is a separate commit to remove the broadcast from the system
server time zone detection logic.

Bug: 140712361
Test: treehugger
Change-Id: I5a84580620dc74c1ab1f73a56379c2256f13ef0e