Expose PreciseDataConnectionState to Public API
1) Use PreciseDataConnectionState in TelephonyRegistry

This is a preparatory CL that is part of cleanup and
removal of a number of unused paramters. This piece
pre-emptively uses PreciseDataConnectionState to
pipe through the apn (name) and the LinkProperties,
which are both part of PreciseDataConnectionState.

2) Send PreciseDataConnectionState to TelephonyRegistry

Build and send PreciseDataConnectionState from Phone
to TelephonyRegistry, which contains all the parameters
that are needed.

3) Update UnitTests for PreciseDataConnectionState

Update unit tests to ensure that DC fail invokes
the correct callback on TelephonyRegistryManager.

4) Test That TelephonyRegistry Caches APN States

Add a test that verifies TelephonyRegistry is tracking
the PreciseDataConnectionState for each APN.

Bug: 143187065
Test: atest TelephonyRegistryTest
Merged-In: I8445808a0cc7a28092004d46ff15d79b661df0b8
Change-Id: I8445808a0cc7a28092004d46ff15d79b661df0b8
(cherry picked from commit 3d245131d66b563f7a497effd7c5ea7ddc7c1406)