Sort quality by timestamp
We can have duration calculation issues if the modem reports call
quality out of order (which it seems to sometimes do).

Instead of relying on the modem, since we don't need the total duration
calculations until end of call, we can just wait until the end and then
calculate duration with all of the quality information sorted by
timestamp.

Bug: 141769213
Test: atest CallQualityMetricsTest
Change-Id: I753d824f5bf518faf2da773c0d9510508e360bd1