Do not set brandOverride if opNames are invalid
This reverts behavior which was changed in ag/7203966 (see line 2185 of
the original SST; when the logic to initialize brandOverride was moved
out of the if-block, it changed the behavior).

This code matches old behavior, which assumes that whenever the device
is not registered, the modem returns an invalid operator.

Bug: 144815455
Test: atest ServiceStateTrackerTest
Change-Id: I415b0fa7f3df85df9f0959d4f546e56b6bf2a2ef