Fix testBroadcastUndeliveredUserLocked
The old code which verifies that one receiver is registered does not
count the receiver that GsmInboundSmsHandler creates when the device is
in TEST_MODE (defaults to on, on userdebug). So we replace the times(1)
call with an atLeastOnce().

Bug: 145232009
Test: atest GsmInboundSmsHandlerTest#testBroadcastUndeliveredUserLocked
Change-Id: Id8ed77afba10fee2d8bca6ddffeb8bd3ab26c0e8
Merged-In: Id8ed77afba10fee2d8bca6ddffeb8bd3ab26c0e8