Update patch set 2
Patch Set 2: Code-Review+1

That definitely fixes the above mentioned crash and is 1:1 in caf like that. I don't think that the original commit which intruduced that in caf is applicable to our repo and to me it seems like this change is squashed into the original commit in caf even though it's unrelated to what the original commit message says.

Patch-set: 2
Reviewer: Gerrit User 21290 <21290@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=+1