Use the new NetworkAgent API
NetworkAgent#updateNetworkInfo is not public or system API, but is
replaced by some new calls. This implements the smallest possible
change converting from the existing objects to the new ones.

This is not completely a no-op change. In particular :
- The network agent used to start in the CONNECTED state, but now
  starts in CONNECTING state and goes to CONNECTED immediately.
- If updates to the extra info, legacy subtype and/or disconnection
  happen concurrently, they get split into multiple calls to
  ConnectivityService, while they used to be a single atomic update.
Both of these changes are supposed to be benign.

Bug: 138306002
Test: FrameworksNetTests FrameworksTelephonyTests
Change-Id: I1ce0209824dc74f38e2223ccfbf5e1c031404f69
Merged-In: I1ce0209824dc74f38e2223ccfbf5e1c031404f69
(clean cherry-pick from internal branch)