Add jarjar rules for telephony tests
The rules combine the base telephony jarjar rules, and test-only jarjar
rules. Test-only rules take precedence as jarjar only applies the first
matching rule to any class.

Includes a rule to rename the NetworkFactory in services.jar, as the
base jarjar rules would otherwise attempt to rename it to the same
package as the one in telephony-common, and cause conflicts.

This is only necessary for telephony tests because telephony has jarjar
rules for NetworkFactory, because it puts telephony-common on the
bootclasspath.

Test: atest FrameworksTelephonyTests
Bug: 138306002
Change-Id: Ia7b80233a2934b285beb4684c3cac2e32521f6d4