Fix to set wrong WFC mode when SIM hot swap
ImsManager#setWfcMode() is called with current WFC mode to update
modem's config when roaming state is changed. When hot swap
home SIM and roaming SIM, the roaming state will be changed.
However, the carrier config may not be loaded in this case. So, wrong
WFC mode may be set.

To fix this problem, prevent set WFC mode state if carrier config
is not loaded.

Test: auto - Passed ImsPhoneTest.
Bug: 151394958

Change-Id: Ie127e3ce33831fc9e9c94303f68923669f635afa