Update Carrier Privileged UIDs when SIM state changes.
CarrierPrivilegesTracker monitors when the SIM state changes. In the
event that the SIM state changes to not present, the SIM-loaded carrier
privilege certs should be cleared and the administrator UIDs should be
refreshed. In the event that the SIM state changes to present, the
SIM-loaded carrier privilege certs should be retrieved and the carrier
privileged UIDs should be refreshed.

Bug: 147391402
Test: Tested on crosshatch build
Change-Id: Ifafb07d19c3f4272610f484d45c268e77e2994eb