Prevent non-cellular request additions
* Non-cellular data connection requests were unnecessarily
being add to a list ordered by apn priority within
PhoneSwitcher.
* This commit prevents non-cellular network requests,
or rather, requests that do not map to any APN from
going into that list.
* We are also checking for an invalid case where a
TelephonySpecifier is used on a network request that does
not map to an APN.  That is being logged as a warning.

Bug: 148225839
Test: DcRequestTest
Test: Tested internet connected to cellular with wifi on and off, viewed logs
Merged-In: Ia813ed8f87e4c60287036c58928e1c99a8d8430c
Change-Id: I34081a443494b14955cc8dd8bc8b4105eb17625b