Update telephony dev ID tests to use new SystemAPI
Based on feedback during the API review of the new SystemAPI for
telephony to check device identifier access the method was moved
from DevicePolicyManager to PermissionManager. This commit updates
the telephony dev ID tests to mock this new check and removes
the tests for checks that are no longer controlled by
TelephonyPermissions.

Bug: 147761267
Test: atest TelephonyPermissionsTest
Test: atest PhoneSubInfoControllerTest
Test: atest SubscriptionControllerTest
Change-Id: I69d434eec22ad8c3cb07eff3a6395e02c88f96ae