DeviceStateMonitor: turn double negative into forward positive logic
On determining when we should turn on/off indication filters,
we have a double negative logic. eg. shouldTurnOff return false
to indicate we should turn on the filter. This is a bit confusing
and prone to introducing bugs.

This CL turns this into forward logic, eg. shouldTurnOn return true
to indicate we should turn on, making the logic a bit easier to
understand.

Bug: 151116827
Test: atest DeviceStateMonitorTest
Change-Id: I9ace326defc9fbf9f14275e0f226abbaae4daa42