Repoint Ims.IsAvailability from tele-cache to ImsService
There was a race case that occurs when an ImsService changes its capability status
and with a call against IsAvailable(capability) within the telephony component which
checks a cache.  This change moves the call in ImsPhoneCallTracker from its internal
cache to instead check the capability status directly within the ImsService.

Bug: 149780842
Test: android.telephony.ims.cts.ImsServiceTest
Change-Id: I145af30c4cc740a7abb20df570b7e6a7d30812f6