Sanitize number from SubInfo if caller does not have READ_PHONE_NUMBERS
getLine1Number has been moved out of the READ_PHONE_STATE permission,
but callers with READ_PHONE_STATE can still query SubscriptionInfo
which contains the number for the subscription. This change will
sanitize the number from the resulting SubscriptionInfo object(s)
if the caller does not meet the requirements to access phone numbers.

Bug: 152063877
Test: atest SubscriptionControllerTest
Change-Id: I5a8823c043d66eaea0460cd3bcdb70a2b6b17716