Remove redundant SubInfo list identifier access checks
To determine if identifiers should be removed from the resulting list
in SubscriptionController#getSubscriptionInfoListFromCacheHelper a
READ_PHONE_STATE and identifier access check is performed up front to
determine if the cached list can be returned as is. However most apps
will fail the identifier access check since a valid subscription ID
cannot be provided up front. This failure adds additional overhead
to the majority of calls and the log entry for the failure can be
confusing as to why a carrier privileged app would fail this check.
These up front checks are removed and the existing READ_PHONE_STATE
and identifier access checks using the subId of the SubscriptionInfo
object are used to build the list and strip out any identifiers if the
caller does not meet the access requirements.

Bug: 152117976
Test: SubscriptionControllerTest
Change-Id: Id8343019a347d2d797851b14555a5c00f2a02f7f