Cache the emergency number list indication
Telephony PhoneFactory registrates RIL via HAL before constructing various components. This can cause a delay for those components to registrate the radio indication they are interested, and thus fail to receive the indication.

To fix that, we need to have a cache and send it out when the corresponding component is registrating it.

Test: manual
Bug: 152818455
Change-Id: Ide3713f0d9314e07af7202c22d2360e7f3b346f7