Implement Radio HAL Fallback Compatibility.
Refer to go/radio-compat. This CL provides a mechanism for partially
implemented HAL versions in Android' Radio HAl, starting with support
for methods in Radio HAL 1.5 to fall back to 1.4 as needed.

Then radio response of a HAL request reports unsupported, RIL will
retry with a new request on the next HAL version and cache the
sucessful request/HAL-version for further handling.

Bug: 151106728
Test: atest com.android.internal.telephony.RILTest

Change-Id: Ied80fccba8b66194aa9c1aa03107c26a49abbd18