Add support for ignoring CEP on conference peer.
Normally, CEP data received on the peer is accepted without question.
Some carriers, however, pass this data even though it's not desired to show
the conference state on the peer.

This change adds support for a new carrier config option to indicate if
the peer data should be used.

Also this CL enhances support for CEP testing on the device.
TelephonyTester supports manual injection of CEP data into an IMS call.
In the past this was done by injecting through the ImsCall.  This, however,
is problematic because there is a telephony command which can be used to
disable CEP for test purposes:
adb shell cmd phone ims conference-event-package disable

Since CEP is disabled in ImsPhoneCallTracker, injecting the test CEP
data through the test IMS call won't work.  This change makes it possible
to inject the CEP data either via the ImsCall (existing) or through
the ImsPhoneCallTracker to ensure it happens even if the override has
disabled it.

This ensures that in manual testing you can disable CEP and still inject
test CEP data onto the device.

Test: Manual test with CEP injection intents.
Test: Added new ImsPhoneCallTracker tests for the new carrier config.
Bug: 150860282
Fixes: 154245604
Merged-In: I3957531afc4bc8adc2c27254db78fec753864bfc
Change-Id: I5135a3378299974df0eafbfaac9d826256af5f7d