Update ImsPhoneCallTracker state on SRVCC complete.
The state of the ImsPhoneCallTracker is cached (like it is in the GSM/CDMA
tracker) in an instance variable, mState.  When calls undergo SRVCC that
state was not being updated back to IDLE, meaning the tracker would report
being off hook after the SRVCC until another call operation takes place
on the tracker.

This bug fixes that by ensuring we call updatePhoneState() when SRVCC
is complete.

Test: Manual test by setting network to 3G and placing a fake emergency
call to verify SRVCC completes and the phone state updates as required.
Test: Added new unit test to confirm update.
Fixes: 154196510

Change-Id: I95f10906acefda6872f345c2f4818cac46df5e75