Fix concurrency issues in ImsPhoneCall.
The class had a couple places where getFirstConnection() was referenced,
but it was possible for concurrent access to take place on the connection
between when getFirstConnection() was first called and when the return
value from it was referenced.  This was easy to fix by just storing the
instance of ImsPhoneConnection returned.

Fixes: 156168243
Test: Manual regression test for IMS phone calls.
Test: Add unit tests for get methods in question; can't really test the
concurrency in the test but can test basic operation.

Change-Id: I341fabf6e7cde869a7a8075f18ed15b593206471