Fix LTE_CA bit to be enabled for < 1.5
Fix a mistake where LTE_CA support was accidentally
broken for devices that didn't support HAL 1.5 but
enabled for devices that do support 1.5. This is
backwards and unintentional.

The previous fix on this bug masked the visible symptoms
of this bug for all devices supporting HAL 1.2 or greater;
however, from the API, this was still wrong, and the bug
still visibly manifests on devices supporting HAL 1.1 or
1.0.

Bug: 154060784
Test: manual
Change-Id: If37bd0b7d6d02e373cd99c478fcd02e948d6d03b