Move pullers to binder thread
Pass direct executor to setPullAtomCallback. This cases the pullers to
run directly on the binder thread, as opposed to the background thread,
which could be busy with other work.

The move to the binder thread means that the pullers could execute
concurrently, but it looks like the pullers have already been made to be
thread-safe.

Test: m
Bug: 157186182
Change-Id: I1513ddb564f181a7c8810cd2e0c97db30e5ca7d5