Fix issue where disconnect cause notification does not fire.
TelephonyRegistry was not firing the notifyDisconnectCause and hence
LISTEN_PRECISE_DATA_CONNECTION_STATE listeners were not getting notified
of the disconnect cause.

The root cause turns out to be that sub id and phone id were swapped in
the parameter list.  That'll do it.

Test: Manual testing to verify registry is updated now.
Test: Update now-failing unit test and modify parameters to make it more
clear which is the phone id and which is the sub id.  The old test was
passing and it was clearly wrong.
Fixes: 157729584
Change-Id: I04a837cbf7447074467ef91804c07c96d5c461dc