SubscriptionController: Don't call getUriForSubscriptionId with invalid ID
It seems like the SIM card can take longer to load on certain legacy CDMA devices,
thus causing the subscription ID not to be valid during early telephony init.

Just bailout to avoid any calls to SubscriptionManager#getUriForSubscriptionId with
an invalid subId, which then makes the phone process crash. Once the SIM is loaded,
SubscriptionInfoUpdater will take care of updating the DB with the proper URI.

Change-Id: I8bd71127aecff9577bb0d3b9d7cab7a01a076e87