ImsServiceController should properly unbind when remote crashes
There is currently an issue where unbind() calls will be ignored
in ImsServiceController because we were mistakenly clearing the
ImsServiceConnection in the onServiceDisconnected method. When in
this state, the ImsServiceController is never properly unbound.

1) Unless unbindService() is explicitly called, the ImsServiceConnection
should not be cleared.

2) Fix some iffy locking issues.

Bug: 159047328
Test: atest FrameworksTelephonyTests CtsTelephonyTestCases:ImsServiceTest
Merged-In: I88d4a5aad637f76b77f056dd6b5a2b5cfd185c87
Change-Id: I88d4a5aad637f76b77f056dd6b5a2b5cfd185c87