Setup TelephonyPermissions mocks for all SubscriptionController tests
The WRITE_SMS appop check in TelephonyPermission's phone number access
verification has been moved after the READ_PHONE_STATE and carrier
privilege check. Due to the new phone number access requirements for
target R+ any tests that query for SubscriptionInfo objects must setup
the mocks for the TelephonyPermission checks; this commit does this in
the setUp method as a majority of the SubscriptionController tests will
perform this query.

Bug: 157642567
Test: atest SubscriptionControllerTest
Test: atest TelephonyPermissionsTest
Test: atest PhoneSubInfoControllerTest
Change-Id: I9e5f0bef0c9c9d820492d92baa48e110abf4c517