Avoid EuiccController#dump off of main thread
Calling EuiccController#dump on main thread will cause a dead lock if
called on the main thread. For example calling EuiccController#dump on
the main thread will result in the following sequence:
1. EuiccController#dump will block on the main thread and wait for
EuiccConnector to bind to LPA
2. EuiccConnector needs to wait for callback `onServiceConnected` which
only happens on the main thread.
3. Resulting in a dead lock as the main thread is already blocked
waiting.

From looking at the bug report, TelephonyDebugService is calling the dump function from main thread. Also the system already calls EuiccController#dump, so the log are duplicated. The fix here will be to remove the #dump from PhoneFactory as it is already called from the system.

Bug: 157838871
Bug: 159765216
Test: Manually tested by `adb bugreport` and see that EuiccController#dump logs are still capture even though we're not on main thread

Change-Id: I9021a3cad04c727749cc7294634fc5839f40bf1d