Change eq method for connection reuse w\ dun apn
When a new sim card is loaded that was previously used, the same apn
settings are reloaded in the Carriers table with different database
ids.  These new database ids caused an issue when DcTracker looks for
existing data connections to use for the dun apn.  This CL changes
the equality check used in that method to the same one used when
cleaning up active data connections which ignores the db id, ignores
the network type bitmask, and only matches on the the relevant
roaming \ non-roaming protocol.

Bug: 158908392
Test: Unit tests
Test: Added DcTrackerTest#testCheckForCompatibleDataConnectionWithDunWhenIdsChange
Change-Id: Iaf8df512e5db35b4dad613b491b718648ce3da11