Fixed time zone issue after out of service
When device becomes out of service, it's not guaranteed
that the device will get a NITZ message again when it camps
on network again.

Removed the calling of handleNetworkUnavailable when device
becomes out of service so the NITZ state machine can still rely
on the last known NITZ data.

Also removed callings of handleNetworkAvailable for symmetry, and
also because it is a no-op method now.

Fix: 159348558
Test: FrameworksTelephonyTests
Change-Id: Icd6fcb6c440ed42ce390eb3c0ec00f0abfcdfd84