Fix a bug where the subtype is not initialized correctly
Note that the initial subtype is still not set correctly at
register() time because there is no way to communicate this
information. This only affects callers that look at the
extra info before the agent is marked connected, so this
should only be internal callers, and they don't actually
rely on this information.

We could fix this by adding a hidden method in NetworkAgentConfig
but this is not strictly necessary, so at this time do the
simplest change only.

Test: FrameworksTelephonyTest
Bug: 161653721
Bug: 161737783
Change-Id: I397d773ac9bb5532d72036d3cb3d37e52c79ee4e
(cherry picked from commit 7403bd5462d052b38809297e2668d224c29a8d0a)