Abort the scan when modem becomes unavailable
The modem may temporarily become unavailable if the radio proxy dies
while a scan is in progress. When this happens, further indications
related to the network scan are not received at the telephony layer,
and subsequently not passed on to Settings. This leads to the scan
being shown as going on indefinitely on the UI.

Fix this by detecting when the modem becomes unavailable and cancelling
the scan.

Bug: 161500625
Test: Manually tested:
    1. Start scan
    2. Kill ril daemon
    3. Verify that the UI shows an error message
Change-Id: If0047e2471952cb49cbe49659d073b83772ac388