Get ISipTransport + ImsService implementation capabilities
Retrieve the capabilities of each ImsService implementation as well
as the ISipTransport implementation if it exists.

If the ImsService implements MMTEL+RCS, it declares the
CAPABILITY_SIP_DELEGATE_CREATION capability, and exposes an
implementation to ISipTransport, the capability will be passed to
listeners to be used for SipDelegate creation.

Test: atest FrameworksTelephonyTests CtsTelephonyTestCases:ImsServiceTest
Merged-In: Iacc798d45bde2140c0bd653791d614280bb25e47
Change-Id: Iacc798d45bde2140c0bd653791d614280bb25e47