Move broadcast on CSS changed outside of GSM check
If the CSS indicator changes, broadcast the change regardless of if the
device is GSM or CDMA. This ensures apps listening to data connection
changes always get the correct state.
notifyAllActiveDataConnections was removed in ag/12289035, so this
change only needs to go into rvc-qpr-dev. Other branches have the fix
in ag/12723538 instead.

Test: atest ServiceStateTrackerTest
Bug: 168529378
Change-Id: Ic0ac32455de98977f0a47d310792767943d07f9d
Merged-In: I39abe0bde114593c2a8cf19d27ae077815c42af3