Ensure IMS calls are disconnected when WPS call is forced over CS.
When a WPS call is placed on the CS domain, ensure that we hangup any
ongoing calls on the IMS domain first.

Test: Added unit test to verify ImsPhoneCallTracker is asked to disconnect
all ongoing calls when a WPS call is placed over CS.
Test: Added unit test to verify ImsPhoneCallTracker is not asked to
disconnect all ongoing calls when a WPS call is placed over IMS.
Test: Manually overrode carrier config to ensure WPS calls go over CS,
started on device IMS call to another number.  Then attempted to place a
WPS call.  Verified that the ongoing IMS call disconnects and the WPS
call is placed.
Bug: 172695501

Change-Id: I7ebb967a8ed6af8dc040cb88c1562bfd32a0e46d