Root locale need a configuration context
KEY_WFC_SPN_USE_ROOT_LOCALE does not return the root
resource in CarrierDisplayNameResolver. The reason is
that setLocale() on default context does not change the
locale in the configuration being used to get resources
in native AssetManager.

For temporary changes to a configuration there is already
a template, you create a new context with new resources.

Test: Manual
Test: atest CarrierDisplayNameResolverTest
Bug: 174237107
Change-Id: I8694685f4d210b32fc7cc3eb8a7f3a79f385ee48