Respond with STATUS_REPORT_STATUS_OK for IMS SMS reports
If a status report format is not same with a sent message, it can not be
handled correctly even if N/W resends same reports again. So a platform
will respond with STATUS_REPORT_STATUS_OK same as CS case always.
STATUS_REPORT_STATUS_ERROR is used only if the platform is not
available.

This scheme is useful when we lose a list for delivery reports due to
restarting the phone process(reboot or crash) as well.

Bug: 174287385
Test: atest ImsSmsDispatcherTest

Signed-off-by: Taesu Lee <taesu82.lee@samsung.com>
Change-Id: Ic2d1c2570894a3ab7afb1feb03388aa3af75880d